Easy fix for "nbf > exp" testing errors
New testing key with an expiration time of ~10 years from the date the key was
generated(2015-04-15).  The delay in this commit is due to lunatic laptop
issues on my end.

Also include the private test key so that we can generate keys on the fly in a
better test later.